refactor(reload): move TileReloaders into its own module
